test(test-agent): move /proxy unit tests to a separate tests file
Rust tests must live under a tests/ directory wired via `mod`, not in an
inline `#[cfg(test)] mod tests` block (the daemon's check-inline-tests
gate enforces this). Relocate the proxy request-parsing tests to
client/tests/proxy.rs alongside the existing ping tests, and expose the
default-target constant to the crate so the test can reference it.
