+++ source/daemon/crates/wardnet-test-agent/src/client/tests/mod.rs
@@ -1,1 +1,2 @@
 mod ping;
+mod proxy;
